Remove PBR split regressions from coverage PR
Revert the PBR runtime chunk splits that grew optional-feature scenes. Keep the production coverage tool and bundle measurement stability fix without changing per-scene bundle sizes.
